+++ project/src/pages/login-page/login-page.tsx
@@ -3,7 +3,7 @@ import { useAppDispatch, useAppSelector} from '../../hooks';
 import { AppRoute, Location } from '../../const';
 import { Link, Navigate } from 'react-router-dom';
 import LoginForm from '../../components/login-form/login-form';
-import { getIsAuthorized } from '../../store/user-process/selectors';
+import { getIsAuthorized } from '../../store/user-data/selectors';
 import { changeLocation } from '../../store/app-process/app-process';
 import { getRandomInt } from '../../utils/utils';
